[refactor] #25 컴포넌트 패키지로 옮기고 변수명 변경
